fixed some bugs and completed basic crud requests

+++ Transaction-API/Transaction-API/Migrations/TransactionDbContextModelSnapshot.cs
@@ -20,9 +20,9 @@ namespace Transaction_API.Migrations
 
             modelBuilder.Entity("Transaction_API.Models.Transaction", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int")
+                        .HasColumnType("bigint")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<decimal>("Amount")
